+++ changelog.md
@@ -2,6 +2,7 @@
 - Added methods for getting and setting lists of specific types for plugin configs
 - Added new methods for configuration files
 - Added empty character constant and methods for colorizing chat messages
+- Added tools for managing isometric objects
 - Changed the logic for saving and loading the standard configuration file
 - Fixed API provisions for chat